refactor: 예외처리 객체 디렉토리 이동, 테스트용 PostException 추가 (#21)
