+++ module-common/build.gradle
@@ -4,6 +4,7 @@ plugins {
 
 dependencies {
     implementation 'org.apache.commons:commons-lang3:3.12.0'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
 }
 
 bootJar {
