+++ module-common/src/main/java/com/back2basics/response/global/result/ResponseCode.java
@@ -1,4 +1,4 @@
-package com.back2basics.response.global;
+package com.back2basics.response.global.result;
 
 public interface ResponseCode {
