+++ module-common/src/main/java/com/back2basics/response/global/result/ResultResponse.java
@@ -1,4 +1,4 @@
-package com.back2basics.response.global;
+package com.back2basics.response.global.result;
 
 import lombok.Getter;
