+++ module-core/build.gradle
@@ -2,6 +2,11 @@ plugins {
     id 'org.springframework.boot' version '3.3.0'
 }
 
+dependencies {
+
+    // module
+    implementation project(':module-common')
+}
 
 bootJar {
     enabled = false
