+++ module-core/src/main/java/com/back2basics/service/post/exception/PostException.java
@@ -1,0 +1,16 @@
+package com.back2basics.service.post.exception;
+
+import com.back2basics.response.global.error.ErrorResponse.FieldError;
+import com.back2basics.response.global.error.exception.CustomException;
+import java.util.List;
+
+public class PostException extends CustomException {
+
+    public PostException(PostErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    public PostException(PostErrorCode errorCode, List<FieldError> errors) {
+        super(errorCode, errors);
+    }
+}
